requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2023.3
requests==2.26.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
urllib3==1.26.16
